requirements.txt
pydantic>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
groq>=0.4.0
